requirements/_base.txt
asgiref==3.2.7
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.44.0
django-crispy-forms==1.9.0
django-heroku==0.3.1
djangorestframework==3.10.3
docutils==0.15.2
gunicorn==19.9.0
idna==2.10
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.2
whitenoise==4.1.4
